Extract ul and li elements into Tiles component
