+++ packages/components/src/index.js
@@ -55,6 +55,7 @@ export { default as Spinner } from './spinner';
 export { default as TabPanel } from './tab-panel';
 export { default as TextControl } from './text-control';
 export { default as TextareaControl } from './textarea-control';
+export { default as Tiles } from './tiles';
 export { default as Tip } from './tip';
 export { default as ToggleControl } from './toggle-control';
 export { default as Toolbar } from './toolbar';
